remove MPIData
