+++ src/pmc_simulation.jl
@@ -24,7 +24,7 @@ function _set_up_starting_vectors(
     start_at, n_replicas, n_spectral, style, initiator, threading, copy_vectors
 )
     if start_at isa AbstractMatrix && size(start_at) == (n_replicas, n_spectral)
-        if eltype(start_at) <: Union{AbstractDVec, RMPI.MPIData} # already dvecs
+        if eltype(start_at) <: AbstractDVec # already dvecs
             if copy_vectors
                 return SMatrix{n_spectral, n_replicas}(deepcopy(v) for v in start_at)
             else
@@ -38,7 +38,7 @@ function _set_up_starting_vectors(
     end
     n_spectral == 1 || throw(ArgumentError("The number of spectral states must match the number of starting vectors."))
     if start_at isa AbstractVector && length(start_at) == n_replicas
-        if eltype(start_at) <: Union{AbstractDVec, RMPI.MPIData} # already dvecs
+        if eltype(start_at) <: AbstractDVec # already dvecs
             if copy_vectors
                 return SMatrix{n_spectral, n_replicas}(deepcopy(v) for v in start_at)
             else
@@ -49,7 +49,7 @@ function _set_up_starting_vectors(
                 default_starting_vector(sa; style, initiator, threading) for sa in start_at
             )
         end
-    elseif start_at isa Union{AbstractDVec, RMPI.MPIData}
+    elseif start_at isa AbstractDVec
         if n_replicas == 1 && !copy_vectors
             return SMatrix{n_spectral, n_replicas}(start_at)
         else
